Deduplicate Medusa runner wrappers
